day01/requirements.txt
asgiref==3.8.1
Django==5.0.8
django-filter==24.3
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
